Drop the tag helper tests and trim comments
Removes the discourseTagText unit tests, folding the pure helpers back into
tagNodeHandler.ts since the separate module existed only so `node:test` could
import them without resolving `obsidian`. That also drops the `test:unit`
script, leaving package.json with just the `@codemirror/language` dependency.

Cuts the remaining comments to the non-obvious rationale only, and removes the
section banners.
